% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_geom_triangles_size_legend.R
\name{draw_geom_triangles_height_legend}
\alias{draw_geom_triangles_height_legend}
\title{Draw geom_triangle legend based on its height (z)}
\usage{
draw_geom_triangles_height_legend(
  width = 1,
  width_scale = 0.1,
  height_scale = 0.1,
  z_values = 1:3,
  n.breaks = 3,
  labels = c("low", "medium", "high"),
  color = "black",
  fill = "black"
)
}
\description{
Draw geom_triangle legend based on its height (z)
}
\examples{

draw_geom_triangles_height_legend()

draw_geom_triangles_height_legend(
width = c(min(mtcars$cyl),median(mtcars$cyl), max(mtcars$cyl)),
labels = c(min(mtcars$cyl),median(mtcars$cyl), max(mtcars$cyl)),
) +
ggtitle("Cylinders") +
theme(plot.title = element_text(hjust = 0.5))

}
