% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_geom_triangles_size_legend.R
\name{draw_geom_triangles_size_legend}
\alias{draw_geom_triangles_size_legend}
\title{Draw geom_triangle legend based on size}
\usage{
draw_geom_triangles_size_legend(
  width = 1/15,
  height_scale = 1/15,
  z_values = 1:3,
  n.breaks = 3,
  labels = c("low", "medium", "high"),
  color = "black",
  fill = "black"
)
}
\description{
Draw geom_triangle legend based on size
}
\examples{

 vertical = TRUE
 width = 1
 height_scale = 1
 z_values = 1:3/3
 n.breaks = 3
 labels = round(1:3/3, 1)


ggplot(
  data = data.frame(x = rep(0, times = n.breaks),
                    y = seq(1,n.breaks),
                    z = quantile(z_values, seq(0, 1, length.out = n.breaks)) \%>\% as.vector(),
                    label = labels
  ),
  mapping = aes(x = x, y = y, z = z, label = label)
) +
geom_triangles()

ggplot(mtcars, aes(x = mpg, y = cyl, z = disp, label = disp)) +
  ggtriangles::geom_triangles(height_scale = 0.001)

}
