% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/draw_geom_triangles_size_legend.R
\name{draw_geom_triangles_width_legend}
\alias{draw_geom_triangles_width_legend}
\title{Draw geom_triangle legend based on its width}
\usage{
draw_geom_triangles_width_legend(
  width = 1:3,
  width_scale = 0.1,
  height_scale = 0.1,
  z_values = 1,
  n.breaks = 3,
  labels = c("low", "medium", "high"),
  color = "black",
  fill = "black"
)
}
\description{
Draw geom_triangle legend based on its width
}
\examples{

draw_geom_triangles_width_legend()

draw_geom_triangles_width_legend(
width = quantile(mtcars$wt, c(.33, .66, 1)),
labels = round(quantile(mtcars$wt, c(.33, .66, 1)), 2)
) +
ggtitle("Weight") +
theme(plot.title = element_text(hjust = 0.5))

}
