% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_triangles.R
\name{geom_triangles}
\alias{geom_triangles}
\title{Triangle plotter -- plot triangles with base x,y and height z}
\usage{
geom_triangles(
  mapping = NULL,
  data = NULL,
  stat = "identity",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\arguments{
\item{x, y}{the x-y coordinates of the midpoint of the base of the triangle}

\item{z}{the height for each triangle, subject to multiplication by height_scale}

\item{width}{the width of the base of the triangle}

\item{height_scale}{the scaling factor (default: 1) for the triangle's height relative to the y values}
}
\description{
Draw triangles with base at (x,y), specifieid width, and height
given by aesthetic argument z.
}
\examples{

# example 1
mtcars \%>\%
ggplot(aes(x = mpg, y = disp, z = cyl, color = hp, fill = hp)) +
  geom_triangles(height_scale = 3) +
  scale_fill_viridis_c() +
  scale_color_viridis_c()

# example 2
iris \%>\%
  ggplot(aes(x = Sepal.Length, y = Sepal.Width, z = Petal.Length)) +
  geom_triangles(width = 0.1, height_scale = 0.05)

# example 3
ggplot(data.frame(x=1:5, y = 1:5, z = (c(5,-.5,3,1.5,-7)/15)), aes(x=x,y=y,z=z)) +
  geom_triangles(width = 0.1)

# example with legend using patchwork
library(patchwork)

z_values <- c(5,-.5,3,1.5,-7)/15

plt <- ggplot(data.frame(x=1:5, y = 1:5, z = z_values), aes(x=x,y=y,z=z)) +
    geom_triangles(width = 0.1)

legend <- draw_geom_triangles_size_legend(z_values = z_values, height_scale = 1.5, width = 0.1)

blank_plot <- ggplot() + theme_void()

(plt + (blank_plot / legend / blank_plot)) +
  plot_layout(ncol = 2, nrow = 1, widths = c(1, .25))

}
\seealso{
GeomTriangle
}
