% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_triangles.R
\docType{data}
\name{GeomTriangles}
\alias{GeomTriangles}
\title{GeomTriangles ggproto}
\format{
An object of class \code{GeomTriangles} (inherits from \code{Geom}, \code{ggproto}, \code{gg}) of length 5.
}
\usage{
GeomTriangles
}
\description{
draw_group creates a "triangle_df" which contains in each row the
coordinates for a vertex of one of the triangles produced.
}
\details{
The simplified version is that given the aes() arguments x, y, z, and
optionally width and height_scale, triangles are drawn which have a
horizontal base which is specified with midpoint (x,y). The vertices adjacent
to the base edge are at points (x-width,y) and (x+width,y).

The 3rd and final vertex is positioned at (x,y+z).

The complicated version is just taking into account the x,y,z and width
values are modified by the x and y coordinate systems and the height_scale
aesthetic argument.

This system is currently designed to plot triangles which have a flat/horizontal
edge with midpoint (x,y), but the system could potentially be extended in future
work to adjust the angle of the triangles.
}
\keyword{datasets}
