% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{ggtriangles}
\alias{ggtriangles}
\alias{ggtriangles-package}
\title{ggtriangles}
\description{
ggtriangles provides the geom_triangles function, a ggplot2 geom layer which will plot the x,y data provided as triangles with height given by argument z.
}
