% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_triangles.R
\name{stat_triangles}
\alias{stat_triangles}
\title{render a polygon layer using the StatTriangles ggproto}
\usage{
stat_triangles(
  mapping = NULL,
  data = NULL,
  geom = "polygon",
  position = "identity",
  na.rm = FALSE,
  show.legend = NA,
  inherit.aes = TRUE,
  ...
)
}
\description{
render a polygon layer using the StatTriangles ggproto
}
\examples{
data.frame(x = 1:5) \%>\%
  ggplot(aes(x = x, y = x, z = x, width = 1)) +
  stat_triangles()
}
