% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom_triangles.R
\docType{data}
\name{StatTriangles}
\alias{StatTriangles}
\title{Stat ggproto that computes a dataframe of triangle vertices given data}
\format{
An object of class \code{StatTriangles} (inherits from \code{Stat}, \code{ggproto}, \code{gg}) of length 3.
}
\usage{
StatTriangles
}
\description{
Given the specified width and height for the triangles, the compute_group
function in this ggproto StatTriangles object create a data.frame (triangle_df
in the code) which has rows which each represent a vertex of a triangle,
with columns for the group (the id of which triangle the vertex belongs to),
x, and y along with other aesthetic parameters like color, alpha, linewidth.
}
\details{
The code supports an angle parameter which allows the user to adjust the angle
the triangles are directed -- however this feature is still under development
and may not work well where the x and y axes of a figure are on significantly
different scales.
}
\keyword{datasets}
